refactor(outfitter): extract scaffold planning internals (OS-438)
